+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 
  - Default settings to simulate the default jetbrains IDEs settings.
  - **Rounded** `action bar` selection
+ - Moved `breadcrumbs` above in the **lower** part of the editor.